README
Target Codevita 2020 : Contest 7

June 25 2020 11:00 AM to June 26 2020 11:00 AM

Description:

These are solutions to the coding problems of the second challenge. Problems are taken from Hackerrank and TCS Codevita challenge. All the codes are compiled at Hackerrank editor and coded in the form of functions. Most of the solutions are built in an easy manner, so that new learners can also understand the code. Codes are well commented.

The link to the problems are given below.

Happy Coding.

https://www.hackerrank.com/contests/target-codevita-7/challenges/bon-appetit

https://www.hackerrank.com/contests/target-codevita-7/challenges/sock-merchant

https://www.hackerrank.com/contests/target-codevita-7/challenges/utopian-tree